+++ model_fetch_firestore/lib/src/model_by_filter/firestore_live_by_filter_bloc.dart
@@ -38,4 +38,9 @@ class FirestoreLiveByFilterBloc<T extends WithId<String>>
     await _subscription?.cancel();
     await super.dispose();
   }
+
+  @override
+  void reload() {
+    // No-op, the model is live.
+  }
 }